akcom-udpecho: making packet timer more accurate
akcom-udpecho was only obtaining the timestamp at the beginning of the main
loop.  However this causes the package timer to be zero if the packet was sent
and the reponse received within the same loop iteration, despite some time
having passed.  The client now obtains the current time at the beginning on the
loop and when processing a recieved packet.

Although this is not a fix for the bug described in Github issue #3, this
likely caused the bug to be triggered when testing the client against a daemon
on localhost.