akcom-udpecho: fix adjust time if delay time is larger than packet time
If the reported delay time is larger than the packet time, this was causing the
adjusted time to roll over and report a signficantly larger time than the
actual packet time.  If the delay time, discard the value as invalid and use
the real packet time instead.

This is one of two possible fixes for Github issue #3.